// rubios-front/src/entities/dto/collections-detail-dto.ts
export interface DetailCollectionDto{
    id: string;
    client: string;
    typeCurrency: string;
    paymentDueDate: string;
    payment: string;
    actualPayment: number;
    commission: string;
    accountabilityDate: string;
    recoveryCommission: boolean;
    paymentDate: string;
    paymentType: string;
    number: number;
}